requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
alpaca-py==0.27.1
alpaca-trade-api==3.2.0
alpha-vantage==2.3.1
annotated-types==0.7.0
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
ccxt==4.3.56
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
duckdb==1.0.0
email_validator==2.2.0
exchange_calendars==4.5.5
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
huggingface-hub==0.23.4
ibapi==9.81.1.post1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpickle==3.2.2
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lumibot==3.5.11
lumiwealth-tradier==0.1.7
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
matplotlib==3.9.0
matplotlib-inline==0.1.7
more-itertools==10.3.0
msgpack==1.0.3
multidict==6.0.5
multitasking==0.0.11
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pandas-datareader==0.10.0
pandas_market_calendars==4.4.1
parso==0.8.4
passlib==1.7.4
peewee==3.17.5
pillow==10.4.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
polygon-api-client==1.14.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==16.1.0
pycares==4.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyluach==2.2.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
Quandl==3.7.0
quantstats-lumi==0.3.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.14.0
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.31
sseclient-py==1.8.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.4.2
termcolor==2.4.0
timedelta==2020.12.3
tokenizers==0.19.1
toolz==0.12.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
uuid==1.30
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
yfinance==0.2.40
